task 10 3

+++ Timmiev_Grigory_lesson_10/task_10_3.py
@@ -31,7 +31,7 @@ class Cell:
     def __truediv__(self, other):
         if not isinstance(other, Cell):
             raise TypeError(f'{other} not a Cell class')
-        return Cell(self.cells / other.cells)
+        return Cell(int(self.cells / other.cells))
 
     def make_order(self, n):
         lines = self.cells // n
